Add Extensions.cs, move AttributeAnalyzer files into Apple dir, add Xamarin.CodeAnalysis.Apple namespace, add public const for IDs, Or for Xamarin.CodeAnalysis.targets file
