+++ src/Xamarin.CodeAnalysis/Xamarin.CodeAnalysis.Apple.targets
@@ -1,5 +1,15 @@
 ﻿<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
 
+  <!--<ItemGroup>
+    <AdditionalFiles Include="@(None)" Condition=" '%(FileName)%(Extension)' == 'Info.plist' ">
+      <Visible>false</Visible>
+    </AdditionalFiles>
+  </ItemGroup>-->
+  
+  <!--<ItemGroup>
+    <AdditionalFiles Include="@(AndroidResource -> WithMetadataValue('RelativeDir', 'Resources\values\'))" />
+  </ItemGroup>-->
+  
   <PropertyGroup>
     <CoreCompileDependsOn>$(CoreCompileDependsOn);_GetInfoPlistAsAdditionalFile</CoreCompileDependsOn>
   </PropertyGroup>
