+++ src/Xamarin.CodeAnalysis/Xamarin.CodeAnalysis.targets
@@ -1,10 +1,10 @@
 <Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
 
   <Import Project="Xamarin.CodeAnalysis.Android.targets" Condition="'$(TargetFrameworkIdentifier)' == 'MonoAndroid'" />
-  <Import Project="Xamarin.CodeAnalysis.MaciOS.targets" Condition="'$(TargetFrameworkIdentifier)' == 'Xamarin.iOS'" />
-  <Import Project="Xamarin.CodeAnalysis.MaciOS.targets" Condition="'$(TargetFrameworkIdentifier)' == 'Xamarin.WatchOS'" />
-  <Import Project="Xamarin.CodeAnalysis.MaciOS.targets" Condition="'$(TargetFrameworkIdentifier)' == 'Xamarin.TVOS'" />
-  <Import Project="Xamarin.CodeAnalysis.MaciOS.targets" Condition="'$(TargetFrameworkIdentifier)' == 'Xamarin.MacOS'" />
+  <Import Project="Xamarin.CodeAnalysis.Apple.targets" Condition="'$(TargetFrameworkIdentifier)' == 'Xamarin.iOS'
+          Or '$(TargetFrameworkIdentifier)' == 'Xamarin.WatchOS'
+          Or '$(TargetFrameworkIdentifier)' == 'Xamarin.TVOS'
+          Or '$(TargetFrameworkIdentifier)' == 'Xamarin.MacOS'" />
 
   <ItemGroup>
     <Analyzer Include="$(MSBuildThisFileDirectory)Xamarin.CodeAnalysis.dll" />
